+++ bpromise.js
@@ -8,22 +8,75 @@
         this.nextPromise = new BPromise();
     }
 
+    // @const
     BPromise.STATUS_PENDING = 'pending';
+    // @const
     BPromise.STATUS_FULFILLED = 'fulfilled';
+    // @const
     BPromise.STATUS_REJECTED = 'rejected';
 
+    // @static
     BPromise.isFunction = function ( f ) {
         // 这么判断是否满足所有浏览器，还需要调研
         return typeof f === 'function';
-    }
-
+    };
+    // @static
+    BPromise.isObject = function ( o ) {
+        return Object.prototype.toString.call( o ) === '[object Object]';
+    };
+    // @static
     BPromise.isValidStatus = function ( s ) {
         return [ BPromise.STATUS_PENDING, BPromise.STATUS_FULFILLED, BPromise.STATUS_REJECTED ].indexOf( s ) > -1;
-    }
+    };
+    // @static
+    BPromise.isPromise = function ( obj ) {
+        return obj.then && obj.getStatus;
+    };
+
+    BPromise.prototype.getFulfilledValue = function () {
+        return this.ffv;
+    };
+
+    BPromise.prototype.getRejectedReason = function () {
+        return this.rjr;
+    };
 
     // promise resolution procdure
     BPromise.prototype.resolvePromise = function ( promise, value ) {
-        // TODO
+
+        if ( promise === value ) throw new TypeError( 'Parameters of resolvePromise can not be the same object.' );
+
+        if ( BPromise.isPromise( value ) ) {
+
+            var valueStatus = value.getStatus();
+            switch( valueStatus ) {
+
+                case BPromise.STATUS_PENDING:
+                    value.then( function ( v ) {
+                        promise.resolve( v );
+                    }, function ( r ) {
+                        promise.reject( r );
+                    } )
+                    break;
+
+                case BPromise.STATUS_FULFILLED:
+                    promise.resolve( value.getFulfilledValue() );
+                    break;
+
+                case BPromise.STATUS_REJECTED:
+                    promise.reject( value.getRejectedReason() );
+                    break;
+            }
+        } else if ( BPromise.isObject( value ) || BPromise.isFunction( value ) ) {
+            // TODO
+        } else {
+            promise.resolve( value );
+        }
+
+    };
+
+    BPromise.prototype.getStatus = function () {
+        return this.status;
     };
 
     BPromise.prototype.resolve = function ( value, promise ) {